# Perf: generate otel traces