For profiling, I raised beyla as daemonset in a local minicube cluster and did load testing on the [microservices-demo](https://github.com/GoogleCloudPlatform/microservices-demo) service.

In the process I found a hot spot: generateTracesWithAttributes. Flame graph showed that 27% of unused memory is used on CopyTo method and about 13% on attrsToMap. I will attach a screenshot of the original profile to the PR. Besides, the profile showed a large amount of memory allocation in AppendEmpty method, but it is inside opentelemetry library and this place will not be optimised (to create `&otlptrace.ScopeSpans{}` it would be better to use arena or sync.Pool, so as not to allocate every time a new variable in heap, it is a big load on cpu because of GC).

```bash
goos: darwin
goarch: arm64
pkg: github.com/grafana/beyla/v2/pkg/export/otel
cpu: Apple M2 Pro
                  │ perf/slow.bench.txt │         perf/fast.bench.txt         │
                  │       sec/op        │   sec/op     vs base                │
GenerateTraces-10           2.868µ ± 1%   2.335µ ± 3%  -18.59% (p=0.000 n=10)

                  │ perf/slow.bench.txt │         perf/fast.bench.txt          │
                  │        B/op         │     B/op      vs base                │
GenerateTraces-10          5.172Ki ± 0%   3.766Ki ± 0%  -27.19% (p=0.000 n=10)

                  │ perf/slow.bench.txt │        perf/fast.bench.txt         │
                  │      allocs/op      │ allocs/op   vs base                │
GenerateTraces-10            70.00 ± 0%   61.00 ± 0%  -12.86% (p=0.000 n=10)
```

![image](https://github.com/user-attachments/assets/9f379bdd-2ac9-413f-ba1f-e989ed151ecb)